pyproject.toml
[project]
name = "ai-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "stagehand>=0.3.0",
    "ipykernel>=6.30.1",
    "temporalio>=1.17.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
